requirements.txt
# Core PDF Processing
pdfplumber==0.11.0
pypdfium2==4.30.0
PyMuPDF==1.24.13
pdf2image==1.17.0

# Image Processing & OCR
Pillow==10.3.0
pytesseract==0.3.10
easyocr==1.7.2

# AI & Machine Learning
openai==1.35.0
anthropic==0.31.0
google-generativeai==0.8.3

# Web & HTML
beautifulsoup4==4.12.3
jinja2==3.1.3
lxml==5.3.0

# Data Processing
pandas==2.2.3
numpy==2.0.2

# Utilities
python-dotenv==1.0.1
tqdm==4.67.1
requests==2.32.3
